Split out server (io) utilities from shared
