+++ dwds/test/events_test.dart
@@ -9,7 +9,7 @@ import 'dart:io';
 import 'package:dwds/src/connections/debug_connection.dart';
 import 'package:dwds/src/events.dart';
 import 'package:dwds/src/services/chrome_proxy_service.dart';
-import 'package:dwds/src/utilities/shared.dart';
+import 'package:dwds/src/utilities/server.dart';
 import 'package:test/test.dart';
 import 'package:vm_service/vm_service.dart';
 import 'package:webdriver/async_core.dart';
